% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remotePull.R
\name{remotePull}
\alias{remotePull}
\title{remotePull}
\usage{
remotePull(files, user, host, to = ".")
}
\arguments{
\item{files}{character vector of file or directory paths to copy.}

\item{user}{user name to use for ssh connection}

\item{host}{host name to use for ssh connection}

\item{to}{directory path to copy the files to}
}
\description{
copy files to a remote host
}
